pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.128.0",
    "httpx>=0.28.1",
    "langchain-chroma>=0.2.1",
    "langchain-openai>=0.3.0",
    "lxml>=6.0.2",
    "openai>=2.15.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",

]
